Shift method processFileText from class Task to Class Storage
