+++ src/main/java/omoh/tasktypes/Task.java
@@ -146,46 +146,6 @@ public class Task {
         return "[" + getStatusIcon() + "] " + description;
     }
 
-    //creates output.txt file to store data
-    //also reads output file
-//    public static void createOrReadOutputFile () {
-//        //code to create new file in data directory
-//        File f = new File("data/output.txt");
-//        try {
-//            Task.readFile();
-//        } catch (FileNotFoundException e) {
-//            System.out.println("File not found");
-//        } catch (EmptyTodoException e) {
-//            System.out.println("Todo task empty");
-//        } catch (CorruptedFileException e) {
-//            System.out.println("Output.txt file format is wrong. File corrupted.");
-//        }
-//    }
-
-    //processes the output.txt file. According to the text it adds tasks and marks tasks
-    public static void processFileText (String[] parts, int iteration) throws EmptyTodoException {
-        String command;
-        switch (parts[0].trim().charAt(0)) {
-        case 'T':
-            command = "todo " + parts[2].trim();
-            Todo.addTodo(command);
-            break;
-        case 'D':
-            command = "deadline " + parts[2].trim() + " /by " + parts[3].trim();
-            Deadline.addDeadline(command);
-            break;
-        case 'E':
-            command = "event " + parts[2].trim() + " /from " + parts[3].trim() + " /to " + parts[4].trim();
-            Event.addEvent(command);
-            break;
-        }
-
-        if (parts[1].trim().equals("1")) {
-            command = "mark " + iteration + 1;
-            modifyDoneStateOrDelete(iteration + 1, command);
-        }
-    }
-
     //Method that prints all the tasks stored in myTaskList array
     public static void printAllTasks() {
         Omoh.printHorizontalLine();
